refactor onboarding screens
